+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [0.5.0]
+
 ### Changed
 
 - Wardens is now published with ESM (`type=module`). It should be backwards compatible.
@@ -75,7 +77,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - `mount`/`unmount` hooks to provision resources
 - `allocate`/`deallocate` for creating hierarchies of resources
 
-[Unreleased]: https://github.com/PsychoLlama/wardens/compare/v0.4.1...HEAD
+[Unreleased]: https://github.com/PsychoLlama/wardens/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/PsychoLlama/wardens/compare/v0.4.1...v0.5.0
 [0.4.1]: https://github.com/PsychoLlama/wardens/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/PsychoLlama/wardens/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/PsychoLlama/wardens/compare/v0.2.0...v0.3.0